Integrate the Unicode data module in to WebURL
