+++ Sources/UnicodeDataStructures/Shared/IndexedTable.swift
