+++ Sources/UnicodeDataStructures/Shared/SegmentedLine.swift
