+++ Tests/UnicodeDataStructuresTests/GenerateData/TableDefinitions/DerivedJoiningType.txt
