+++ Tests/UnicodeDataStructuresTests/GenerateData/TableDefinitions/IdnaMappingTable.txt
