+++ Tests/UnicodeDataStructuresTests/SegmentedLineTests.swift
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import UnicodeDataTools
+import UnicodeDataStructures
 import XCTest
 
 fileprivate func XCTAssertSegments<Bound, Value>(
