--- UnicodeData/Sources/UnicodeDataTools/Shared
@@ -1,1 +1,0 @@
-../../../Sources/IDNA/Shared
